The fourth commit using generic views.
Used the basic generic view system.
These views represent a common case of basic Web development: getting data from the database according to a parameter passed in the URL,
loading a template and returning the rendered template. Because this is so common, Django provides a shortcut, called the “generic views” system.
Generic views abstract common patterns to the point where you don’t even need to write Python code to write an app.